# swagger.yaml
openapi: 3.0.0
info:
  title: BiteBuddies API
  description: API documentation for BiteBuddies application services
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Account Service
  - url: http://localhost:5001
    description: Availability Service
  - url: http://localhost:5002
    description: Restaurant Service
  - url: http://localhost:5006
    description: Transcription Service
  - url: http://localhost:5007
    description: Chatbot Service
  - url: http://localhost:5010
    description: Search Service
...
paths:
...
  /transcriptions:
    get:
      tags:
        - Transcription
      summary: Get transcriptions
      description: Retrieve a list of transcriptions for a user
      responses:
        "200":
          description: List of transcriptions
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: object
                    properties:
                      transcriptions:
                        type: array
                        items:
                          type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
  /upload:
    post:
      tags:
        - Transcription
      summary: Upload audio for transcription
      description: Upload an audio file and get a text transcription
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - audio
              properties:
                audio:
                  type: string
                  format: binary
                  description: Audio file to transcribe
                moduleId:
                  type: string
                  description: Module identifier
                userEmail:
                  type: string
                  description: User's email address
      responses:
        "200":
          description: Transcription result
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: object
                    properties:
                      transcription:
                        type: string
                      transcription_id:
                        type: string
                      file_path:
                        type: string
                      db_saved:
                        type: boolean
                      questions:
                        type: array
                        items:
                          type: object
        "400":
          description: Bad request - no audio file uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
        "404":
          description: No speech detected in audio
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
  /question:
    get:
      tags:
        - Transcription
      summary: Get personalized questions
      description: Generate personalized questions for users based on profiles
      parameters:
        - name: userEmail
          in: query
          required: true
          description: User's email address
          schema:
            type: string
        - name: meetingId
          in: query
          required: true
          description: Meeting identifier
          schema:
            type: string
      responses:
        "200":
          description: List of personalized questions
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  questions:
                    type: array
                    items:
                      type: object
        "400":
          description: Bad request - missing parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
        "404":
          description: Meeting or profile not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string

  # =========== CHATBOT SERVICE ===========
...